requirements.txt
fastapi
uvicorn[standard]
python-dotenv
sqlalchemy
psycopg2-binary
alembic
cryptography
tweepy
requests
spacy
celery[redis]
redis
aiohttp
pydantic
python-multipart
typing-extensions